order gateway

+++ order/src/order.service.ts
@@ -6,15 +6,22 @@ import { IOrderGetResponse, IOrderGetsResponse } from './interfaces/get-order';
 import { IOrderCreateResponse } from './interfaces/create-order';
 import { firstValueFrom } from 'rxjs';
 import { IProductGetResponse } from './interfaces/get-product';
+import { IOrderUpdateResponse } from './interfaces/update-order';
+import { IOrderDeleteResponse } from './interfaces/delete-order';
 
 @Injectable()
 export class OrderService {
-    private client: ClientProxy;
+    private productService: ClientProxy;
+    private userService: ClientProxy;
 
     constructor(private prisma: PrismaService) {
-        this.client = ClientProxyFactory.create({
+        this.productService = ClientProxyFactory.create({
             transport: Transport.TCP,
             options: { port: 3000 }
+        });
+        this.userService = ClientProxyFactory.create({
+            transport: Transport.TCP,
+            options: { port: 3002 }
         });
     }
 
@@ -75,8 +82,16 @@ export class OrderService {
     async createOrder(data: Order): Promise<IOrderCreateResponse> {
         try {
             const { productId, quantity, userId } = data;
-            const prdouct: IProductGetResponse = await firstValueFrom(this.client.send('get-product', productId));
-            if (!prdouct) {
+            const prdouct: IProductGetResponse = await firstValueFrom(this.productService.send('get-product', productId));
+            if ( quantity < 0 ) {
+                return {
+                    status: HttpStatus.BAD_REQUEST,
+                    message: 'Quantity cannot be negative',
+                    data: null,
+                    errors: null
+                }
+            }
+            if (!prdouct.data) {
                 return {
                     status: HttpStatus.NOT_FOUND,
                     message: 'Product not found',
@@ -92,8 +107,8 @@ export class OrderService {
                     errors: null
                 }
             }
-            const user = await firstValueFrom(this.client.send('get-user', userId));
-            if ( !user ) {
+            const user = await firstValueFrom(this.userService.send('get-user', userId));
+            if ( !user.data ) {
                 return {
                     status: HttpStatus.NOT_FOUND,
                     message: 'User not found',
@@ -101,7 +116,7 @@ export class OrderService {
                     errors: null
                 }
             }
-            await firstValueFrom(this.client.send('update-product', { id: productId, data: { stock: prdouct.data.stock - quantity } }));
+            await firstValueFrom(this.productService.send('update-product', { id: productId, data: { stock: prdouct.data.stock - quantity } }));
             
             const order = await this.prisma.order.create({ data });
             if (!order) {
@@ -119,6 +134,7 @@ export class OrderService {
                 errors: null
             }
         } catch (error) {
+            console.log(error);
             return {
                 status: HttpStatus.INTERNAL_SERVER_ERROR,
                 message: 'Internal Server Error',
@@ -128,11 +144,81 @@ export class OrderService {
         }
     }
 
-    async updateOrder(id: number, data: Order) {
-        return this.prisma.order.update({ where: { id }, data });
+    async updateOrder(id: number, data: Order): Promise<IOrderUpdateResponse> {
+        try {
+            const { productId, quantity, userId } = data;
+            if ( quantity < 0 ) {
+                return {
+                    status: HttpStatus.BAD_REQUEST,
+                    message: 'Quantity cannot be negative',
+                    errors: null
+                }
+            }
+            if ( !productId || !quantity || !userId ) {
+                return {
+                    status: HttpStatus.BAD_REQUEST,
+                    message: 'Product, quantity, and user id are required',
+                    errors: null
+                }
+            }
+
+            const product: IProductGetResponse = await firstValueFrom(this.productService.send('get-product', productId));
+            if ( !product.data ) {
+                return {
+                    status: HttpStatus.NOT_FOUND,
+                    message: 'Product not found',
+                    errors: null
+                }
+            }
+            const user = await firstValueFrom(this.userService.send('get-user', userId));
+            if ( !user.data ) {
+                return {
+                    status: HttpStatus.NOT_FOUND,
+                    message: 'User not found',
+                    errors: null
+                }
+            }
+            await this.prisma.order.update({ where: { id }, data });
+            await firstValueFrom(this.productService.send('update-product', { id: productId, data: { stock: { increasement:  product.data.stock - quantity   } } } ));
+            return {
+                status: HttpStatus.OK,
+                message: 'Order has been updated',
+                errors: null
+
+            }
+        } catch (error) {
+            return {
+                status: HttpStatus.INTERNAL_SERVER_ERROR,
+                message: 'Internal Server Error',
+                errors: error
+
+            }
+        }
     }
 
-    async deleteOrder(id: number) {
-        return this.prisma.order.delete({ where: { id } });
+    async deleteOrder(id: number): Promise<IOrderDeleteResponse> {
+        try {
+           const order = await this.prisma.order.findUnique({ where: { id } });
+            if ( !order ) {
+                return {
+                        status: HttpStatus.NOT_FOUND,
+                        message: 'Order not found',
+                        errors: null
+                }
+            } 
+            await this.prisma.order.delete({ where: { id } });
+            await firstValueFrom(this.productService.send('update-product', { id: order.productId, data: { stock: { increasement: order.quantity } } }));
+            return {
+                status: HttpStatus.OK,
+                message: 'Order has been deleted',
+                errors: null
+            }
+        } catch (error) {
+            return {
+                status: HttpStatus.INTERNAL_SERVER_ERROR,
+                message: 'Internal Server Error',
+                errors: error
+            }
+        }
     }
 }
